error handling

+++ emp_management/src/main/java/com/employee_management/services/EmployeeService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.employee_management.entity.Employee;
+import com.employee_management.exception.ResourceNotFoundException;
 import com.employee_management.repository.EmployeeRepository;
 
 @Service
@@ -17,33 +18,70 @@ public class EmployeeService {
 
     // create employees
     public Employee addEmployee(Employee emp) {
+        try {
+            if (emp.getName() == null || emp.getDepartment() == null) {
+                throw new IllegalArgumentException("Name and Department are required fields.");
+            }
+            return empRepository.save(emp);
 
-        return empRepository.save(emp);
+        } catch (Exception e) {
+            // TODO: handle exception
+            System.out.println(e.getMessage());
+            return null;
+        }
+
     }
 
     // get all employee
     public List<Employee> getAllEmployees() {
-        return empRepository.findAll();
+        try {
+            return empRepository.findAll();
+
+        } catch (Exception e) {
+            // TODO: handle exception
+            System.out.println("Error in getting data");
+            return null;
+        }
     }
 
     // get employee by id
     public Optional<Employee> getEmployeeById(Long id) {
-        return empRepository.findById(id);
+        try {
+            if (id == null) {
+                throw new IllegalArgumentException("Employee Id is required");
+            }
+            return empRepository.findById(id);
+
+        } catch (Exception e) {
+            // TODO: handle exception
+            System.out.println(e.getMessage());
+            return null;
+        }
     }
 
     // update employee
-    public Employee updateEmployee(Long id, Employee emp){
-          return empRepository.findById(id).map((el)->{
+    public Employee updateEmployee(Long id, Employee emp) {
+        return empRepository.findById(id).map((el) -> {
             el.setName(emp.getName());
             el.setDepartment(emp.getDepartment());
             el.setSalary(emp.getSalary());
             return empRepository.save(el);
-         }).orElseThrow(()-> new RuntimeException("Employee not found"));
+        }).orElseThrow(() -> new ResourceNotFoundException("Employee with id " + id + "not found"));
     }
 
     // delete employee
     public String deleteEmployee(Long id) {
-        empRepository.deleteById(id);
-        return "Deleted";
+        try {
+            if (id == null) {
+                throw new IllegalArgumentException("Employee Id is required");
+            }
+            empRepository.deleteById(id);
+            return "Deleted";
+        } catch (Exception e) {
+            // TODO: handle exception
+            System.out.println(e.getMessage());
+            return null;
+        }
+
     }
 }
